+++ addon/helpers/deprecated/ago.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import { helperFactory } from '../moment-from-now';
+
+export default helperFactory(function() {
+  Ember.deprecate('ember-moment: `ago` helper has been renamed to `moment-from-now`');
+});