+++ addon/helpers/moment-from-now.js
@@ -1,0 +1,41 @@
+import Ember from 'ember';
+import moment from 'moment';
+
+const { later:runLater } = Ember.run;
+
+export var helperFactory = function(cb) {
+  let ago;
+
+  if (Ember.Helper) {
+    ago = Ember.Helper.extend({
+      compute: function(params, hash) {
+        if (typeof cb === 'function') {
+          cb();
+        }
+        if (params.length === 0) {
+          throw new TypeError('Invalid Number of arguments, expected at least 1');
+        }
+        if (hash.interval) {
+          runLater(this, function() {
+            this.recompute();
+          }, hash.interval);
+        }
+        return moment.apply(this, params).fromNow();
+      }
+    });
+  }
+  else {
+    ago = function(params) {
+      if (typeof cb === 'function') {
+        cb();
+      }
+      if (params.length === 0) {
+        throw new TypeError('Invalid Number of arguments, expected at least 1');
+      }
+      return moment.apply(this, params).fromNow();
+    };
+  }
+  return ago;
+};
+
+export default helperFactory();