+++ app/helpers/ago.js
@@ -1,4 +1,4 @@
-import agoHelper from 'ember-moment/helpers/ago';
+import deprecatedAgo from 'ember-moment/helpers/deprecated/ago';
 import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
 
-export default makeBoundHelper(agoHelper);
+export default makeBoundHelper(deprecatedAgo);