+++ app/helpers/moment-duration.js
@@ -1,4 +1,4 @@
-import durationHelper from 'ember-moment/helpers/deprecated/duration';
+import durationHelper from 'ember-moment/helpers/moment-duration';
 import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
 
 export default makeBoundHelper(durationHelper);