+++ app/helpers/moment-format.js
@@ -1,4 +1,4 @@
-import momentHelper from 'ember-moment/helpers/deprecated/moment';
+import momentHelper from 'ember-moment/helpers/moment-format';
 import makeBoundHelper from 'ember-moment/utils/make-bound-helper';
 
 export default makeBoundHelper(momentHelper);