+++ tests/unit/helpers/moment-format-test.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import moment from 'moment';
 import hbs from 'htmlbars-inline-precompile';
-import momentHelper from 'ember-moment/helpers/moment';
+import momentHelper from 'ember-moment/helpers/moment-format';
 import { moduleFor, test } from 'ember-qunit';
 import date from '../../helpers/date';
 import callHelper from '../../helpers/call-helper';
@@ -9,7 +9,7 @@ import { runAppend, runDestroy } from '../../helpers/run-append';
 
 const FAKE_HANDLEBARS_CONTEXT = {};
 
-moduleFor('helper:moment', {
+moduleFor('helper:moment-format', {
   setup() {
     moment.locale('en');
     this.container.register('view:basic', Ember.View);
@@ -34,12 +34,12 @@ test('three args (date, outputFormat, inputFormat)', (assert) => {
 });
 
 test('change date input and change is reflected by bound helper', function(assert) {
-  let context = Ember.Object.create({
+  const context = Ember.Object.create({
     date: date(0)
   });
 
-  let view = this.container.lookupFactory('view:basic').create({
-    template: hbs`{{moment date}}`,
+  const view = this.container.lookupFactory('view:basic').create({
+    template: hbs`{{moment-format date}}`,
     context: context
   });